move some stuff to `kfilters`
