+++ pkg/security/probe/kfilters/approvers_test.go
@@ -6,7 +6,7 @@
 //go:build linux
 // +build linux
 
-package probe
+package kfilters
 
 import (
 	"testing"
