+++ pkg/security/probe/kfilters/capabilities.go
@@ -6,7 +6,7 @@
 //go:build linux
 // +build linux
 
-package probe
+package kfilters
 
 import (
 	"path"
